requirements.txt
pip==24.0
setuptools==69.5.1
wheel==0.43.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
pg8000==1.31.2
fastapi==0.111.0
Werkzeug==3.0.3
pydantic==2.7.1
cryptography==42.0.7
python-jose==3.3.0
